Use new follow data source update follow data on firebase.
